% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixed}
\alias{fixed}
\title{Fixed Effects SO MA of overlapping studies}
\usage{
fixed(ES, SE, cor_mat)
}
\arguments{
\item{ES}{a numeric vector of effect sizes from the meta analyses}

\item{SE}{a numeric vector of standard errors corresponding to the effect sizes}

\item{cor_mat}{a matrix of correlations between effect sizes}
}
\value{
\item{Mean ES} mean effect size estimate from fixed effects second order meta analysis

\item{SE} standard error of the mean effect size estimate

\item{Q} test statistic of the test for heterogeneity

\item{df} degrees of freedom of the test for heterogeneity

\item{p} p-value of the test for heterogeneity
}
\description{
Fixed Effects SO MA of overlapping studies
}
