% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{random}
\alias{random}
\title{Random Effects SO MA of overlapping studies}
\usage{
random(ES, SE, cor_mat, iter, type)
}
\arguments{
\item{ES}{Effect sizes from the meta analyses}

\item{SE}{Standard errors corresponding to the effect sizes}

\item{cor_mat}{Covariance matrix of the effect sizes}

\item{iter}{Number of iterations}

\item{type}{ML or REML}
}
\value{
SOMA estimated effect size, standard error, Q statistic, degrees of freedom, p-value, and estimated tau-squared.
}
\description{
Random Effects SO MA of overlapping studies
}
