% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixed}
\alias{fixed}
\title{Fixed Effects SO MA of overlapping studies}
\usage{
fixed(ES, SE, cor_mat)
}
\arguments{
\item{ES}{Effect sizes from the meta analyses}

\item{SE}{Standard errors corresponding to the effect sizes}

\item{cor_mat}{Covariance matrix of the effect sizes}
}
\value{
* `Mean ES` mean effect size estimate

* `SE` standard error

* `Q` test statistic of test for heterogeneity

* `df` degrees of freedom of test for heterogeneity

* `p` p-value of test for heterogeneity
}
\description{
Fixed Effects SO MA of overlapping studies
}
