% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixed}
\alias{fixed}
\title{Fixed Effects SO MA of overlapping studies}
\usage{
fixed(ES, SE, cor_mat)
}
\arguments{
\item{ES}{Effect sizes from the meta analyses}

\item{SE}{Standard errors corresponding to the effect sizes}

\item{cor_mat}{Covariance matrix of the effect sizes}
}
\value{
SOMA estimated effect size, standard error, Q statistic, degrees of freedom, and p-value
}
\description{
Fixed Effects SO MA of overlapping studies
}
