% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{random}
\alias{random}
\title{Random effects second order meta analysis}
\usage{
random(ES, SE, cor_mat, iter, type)
}
\arguments{
\item{ES}{Effect sizes from the meta analyses}

\item{SE}{Standard errors corresponding to the effect sizes}

\item{cor_mat}{Covariance matrix of the effect sizes}

\item{iter}{Number of iterations}

\item{type}{ML or REML}
}
\value{
\item{Mean ES} mean effect size estimate from fixed effects second order meta analysis

\item{SE} standard error of the mean effect size estimate

\item{Q} test statistic of the test for heterogeneity

\item{df} degrees of freedom of the test for heterogeneity

\item{p} p-value of the test for heterogeneity

\item{tau} estimate of heterogeneity
}
\description{
Random effects second order meta analysis of overlapping studies.
}
